Move HardDifficultyPanel to common/panels.
