+++ lib/widgets/my_app.dart
@@ -7,9 +7,9 @@ import 'package:bratacha/managers/level_manager.dart';
 import 'package:bratacha/modules/dialog_manager/dialog_manager.dart';
 import 'package:bratacha/modules/player_data/player_data.dart';
 import 'package:bratacha/modules/settings_database/settings_database.dart';
+import 'package:bratacha/widgets/common/panels/hard_difficulty_panel/hard_difficulty_cubit.dart';
 import 'package:bratacha/widgets/game_screen/game_screen.dart';
 import 'package:bratacha/widgets/home_screen/home_screen.dart';
-import 'package:bratacha/widgets/home_screen/settings_tab/hard_difficulty_cubit.dart';
 import 'package:bratacha/widgets/home_screen/settings_tab/language_cubit.dart';
 import 'package:bratacha/widgets/learn_screen/learn_screen.dart';
 import 'package:bratacha/widgets/onboarding_screen/onboarding_screen.dart';
